........
  r8079 | sinbad | 2008-10-07 11:31:02 +0100 (Tue, 07 Oct 2008) | 1 line

  Normalise rotation quaternions to eliminate the progressive drift which can happen in some cases (resulting in skewed matrices)
........
  r8082 | sinbad | 2008-10-07 18:48:47 +0100 (Tue, 07 Oct 2008) | 1 line

  Fix a bug in the new compositor script compiler - the 'clear' and 'stencil' passes are supposed to pick up their settings (buffers, colour_value etc) from nested sections, not attributes directly under the pass. Added new translators to handle these nested areas and moved the attribute parsing into them. This fixes bad parsing of compositors with custom clear / stencil settings that used to work in 1.4.
........
  r8085 | sinbad | 2008-10-08 22:51:55 +0100 (Wed, 08 Oct 2008) | 1 line

  Fix transposed matrix parsing in new script compiler; affected 'transform' attribute in texture_unit and also shader parameters expressed as a matrix in script. Addressing of matrix4 is [row][column] and not [column][row].
........
  r8086 | sinbad | 2008-10-08 23:24:39 +0100 (Wed, 08 Oct 2008) | 1 line

  3rd-party license clarification
........
  r8087 | sinbad | 2008-10-08 23:41:39 +0100 (Wed, 08 Oct 2008) | 1 line

  Fix to InstancedGeometry::destroy - clear list
........
  r8089 | sinbad | 2008-10-10 14:02:36 +0100 (Fri, 10 Oct 2008) | 1 line

  Change SDK script so that uninstall doesn't remove any new nested folders created by the user underneath it.
........
  r8090 | sinbad | 2008-10-10 14:11:55 +0100 (Fri, 10 Oct 2008) | 1 line

  Version bump, remove RC1
........
  r8091 | sinbad | 2008-10-10 18:07:26 +0100 (Fri, 10 Oct 2008) | 1 line

  Fix compile when using (deprecated) GTK platform
........
  r8092 | sinbad | 2008-10-11 15:51:18 +0100 (Sat, 11 Oct 2008) | 1 line

  Patch ID: 2149780 - Patch autotools files for PCZSceneManager and new Samples
........
  r8093 | sinbad | 2008-10-11 15:56:11 +0100 (Sat, 11 Oct 2008) | 1 line

  InstancedGeometry should be looking for the best technique, not the first technique
........
  r8094 | sinbad | 2008-10-11 16:04:16 +0100 (Sat, 11 Oct 2008) | 1 line

  Patch 2158215: Fix for Technique::movePass
........
  r8095 | sinbad | 2008-10-11 16:15:37 +0100 (Sat, 11 Oct 2008) | 2 lines

  Patch 2111245: Camera::getCameraToViewportBoxVolume extra plane generated
  Also added optional far plane inclusion as requested, non interface breaking
........
  r8096 | sinbad | 2008-10-11 16:35:43 +0100 (Sat, 11 Oct 2008) | 1 line

  Extra checking in AutoParamDataSource to prevent accidental out of bounds access to light arrays
........
  r8097 | sinbad | 2008-10-11 16:45:20 +0100 (Sat, 11 Oct 2008) | 1 line

  Protect hash recalculation process from race conditions by adding mutexes to cover the change conditions
........
  r8098 | sinbad | 2008-10-11 17:36:55 +0100 (Sat, 11 Oct 2008) | 2 lines

  Fix plugin loading on OS X - don't add '.' to plugin dir if empty, OS X uses Resources area
........
  r8099 | sinbad | 2008-10-11 17:42:43 +0100 (Sat, 11 Oct 2008) | 1 line

  Patch 2076309 - Mac/Carbon support for "border"
........
  r8100 | sinbad | 2008-10-11 17:47:33 +0100 (Sat, 11 Oct 2008) | 1 line

  Patch 2080660 - Add Mac/Carbon support for new windowClosing event
........
  r8102 | sinbad | 2008-10-12 12:30:06 +0100 (Sun, 12 Oct 2008) | 1 line

  Patch ID: 2149780 - Patch autotools files for PCZSceneManager and new Samples (add missing files)
........
  r8103 | sinbad | 2008-10-12 12:32:06 +0100 (Sun, 12 Oct 2008) | 1 line

  Patch ID: 2149780 - Patch autotools files for PCZSceneManager and new Samples (add missing files)
........
  r8104 | sinbad | 2008-10-12 12:32:41 +0100 (Sun, 12 Oct 2008) | 1 line

  Patch ID: 2149780 - Patch autotools files for PCZSceneManager and new Samples (add missing files)
........
  r8105 | sinbad | 2008-10-12 18:16:52 +0100 (Sun, 12 Oct 2008) | 2 lines

  Rename OgreOctree.h in OctreeZone plugin to avoid name clash with existing file on OctreeSceneManager plugin.
  Add PCZ plugins to Mac OSX build
........
  r8106 | sinbad | 2008-10-12 20:33:46 +0100 (Sun, 12 Oct 2008) | 1 line

  Update dependencies so that PCZ / OctreeZone plugins always get built, to avoid problems with plugins.cfg trying to load them
........
  r8107 | sinbad | 2008-10-12 20:55:23 +0100 (Sun, 12 Oct 2008) | 1 line

  Update dependencies so that PCZ / OctreeZone plugins always get built, to avoid problems with plugins.cfg trying to load them
........
  r8108 | sinbad | 2008-10-12 20:57:46 +0100 (Sun, 12 Oct 2008) | 1 line

  Update dependencies so that PCZ / OctreeZone plugins always get built, to avoid problems with plugins.cfg trying to load them
........
  r8109 | sinbad | 2008-10-15 17:05:35 +0100 (Wed, 15 Oct 2008) | 1 line

  Patch 2168040: fix deadlock in Pass hash update process
........
  r8110 | sinbad | 2008-10-20 15:52:49 +0100 (Mon, 20 Oct 2008) | 1 line

  Limit FOV on standard shadow camera setup so it doesn't approach 180 degrees
........
  r8111 | sinbad | 2008-10-22 15:17:19 +0100 (Wed, 22 Oct 2008) | 2 lines

  Fix 'make install' on Linux, which was failing because of the re-link between OctreeZone and PCZ plugins. Not sure if this is the best way to do it, but it works (giving PCZ a standard 'lib' prefix and adding the library path).
........
  r8112 | bjj1478 | 2008-10-24 15:47:14 +0100 (Fri, 24 Oct 2008) | 1 line

  Added support for leading '/' in tokens.
........
  r8113 | sinbad | 2008-10-26 18:07:37 +0000 (Sun, 26 Oct 2008) | 1 line

  Memory allocator bindings for particlefx plugin
........
  r8114 | sinbad | 2008-10-26 19:02:09 +0000 (Sun, 26 Oct 2008) | 1 line

  Memory allocator bindings for PCZ plugins
........
  r8115 | sinbad | 2008-10-26 19:25:07 +0000 (Sun, 26 Oct 2008) | 1 line

  Fix build when using nedmalloc allocator
........
  r8116 | sinbad | 2008-10-27 11:57:09 +0000 (Mon, 27 Oct 2008) | 3 lines

  Include nedmalloc in source release (Boost license, and convenient since it can be built in)
  Hide nedmalloc implementation inside OgreMain to make linking simpler
  Nedmalloc seems noticeably faster - consider making it the default?
........
  r8117 | sinbad | 2008-10-27 12:01:18 +0000 (Mon, 27 Oct 2008) | 1 line

  Some configurations had old nedmalloc location
........
  r8118 | sinbad | 2008-10-27 13:06:24 +0000 (Mon, 27 Oct 2008) | 2 lines

  Linux build fix for nedmalloc
........
  r8119 | sinbad | 2008-10-27 15:42:57 +0000 (Mon, 27 Oct 2008) | 1 line

  Fix MSVC build when nedmalloc disabled
........
  r8122 | sinbad | 2008-10-29 08:14:13 +0000 (Wed, 29 Oct 2008) | 2 lines

  Update Mac OS X to allow use of nedmalloc
  Entirely remove the StdAllocPolicy when nedmalloc in use.
........
  r8123 | sinbad | 2008-10-30 11:16:44 +0000 (Thu, 30 Oct 2008) | 1 line

  Fix MinGW build
........
  r8124 | sinbad | 2008-10-30 11:30:37 +0000 (Thu, 30 Oct 2008) | 1 line

  Fix problem with setting up a controller on a TextureUnitState when one already exists (e.g. calling setAnimatedTextureName more than once)
........
  r8125 | sinbad | 2008-10-30 12:36:12 +0000 (Thu, 30 Oct 2008) | 1 line

  Slight optimisation to BillboardSet - use Matrix3 instead of Quaternion to orient, uses less instructions.
........
  r8126 | sinbad | 2008-10-30 13:04:38 +0000 (Thu, 30 Oct 2008) | 1 line

  Default preferred render texture mode to FBO now - this appears to be working just fine in Leopard on my July 2007 MBP.
........
  r8127 | sinbad | 2008-10-30 13:42:45 +0000 (Thu, 30 Oct 2008) | 1 line

  Improve LiSPSM quality when camera direction becomes parallel with the light direction (scale up the opt adjust factor automatically)
........
  r8128 | sinbad | 2008-10-30 16:45:34 +0000 (Thu, 30 Oct 2008) | 1 line

  Fix weird darkening of grass blades while camera moves (bug 190)
........
  r8129 | sinbad | 2008-10-30 17:24:29 +0000 (Thu, 30 Oct 2008) | 1 line

  Definitely prefer FBO over PBuffer on OSX now. People will need to delete their ogre.cfg files though.
........

+++ PlugIns/ParticleFX/src/OgreColourImageAffector.cpp
@@ -74,7 +74,7 @@ namespace Ogre {
             _loadImage();
         }
 
-		int				   width			= mColourImage.getWidth()  - 1;
+		int				   width			= (int)mColourImage.getWidth()  - 1;
         
 		while (!pi.end())
 		{
